Fix caml_globals_map duplicate entries
Used `replace` instead of `add` on the interfaces Hash table (which is
being used as a set), to avoid duplicate entries. The duplicate entries
lead to enormous `caml_globals_map` values.

Also use `replace` instead of `add` on another hash table usage, not
because it was causing any problems but because it seems like `add`
should be avoided outside unusual situations.