Don't drop deleted code ids
A deleted code id can still appear in a set of closures. In this case,
we know that the closure won't actually be invoked, but in general we
can't remove it from the set because it might be the only closure and
we can't have a set of zero closures. (A set of closures surviving
despite pointing entirely to deleted code would be surprising, but in
principle it's possible if something manages to use the environment.)

Since it's going to appear in the set of closures, we need the code id
to end up in the exported code. We put in a dummy calling convention,
since it's never going to be called anyway.