Add inlined debuginfo when simplifying cont apps
Failing to add inlined debuginfo meant that inlining a function that
raises an exception would leave the debuginfo in the inlined code
alone, thus leaving out the calling function from the stack trace.

So, the source

    let my_raise exn = raise exn
    let f () = my_raise Not_found

gets CPS'd to

    let my_raise exn k e = raise e exn [dbg my_raise]
    let f () k e = my_raise exn k e [dbg f]

and then inlining _should_ produce

    let f () k e = raise e exn [dbg f;my_raise]

but if we forget to adjust the debuginfo for a continuation
application (including a `raise`), we get instead

    let f () k e = raise e exn [dbg my_raise]

and `f` now doesn't appear in the stack trace.

(Worth noting that "inlined debuginfo" is actually about the
_caller(s)_, not the inlined function.)