Don't rewrite `n = m - c` if `n` or `m` is negative
The identity `n = m - c <=> c = m - n` only holds if `n`, `m`, and `c`
are all the same width. There are corner cases when `n` and `m` get
sign-extended and one of them is negative, most obviously when `n =
min_int` since then `-n = n`. For the moment, we have z3 telling us
that it suffices that `n` and `m` are both non-negative, so this
should keep us safe.