Use Format to print Cfg instructions
Extends the interface to use Format, but does not change the formatting,
which leaves much to be desired and will be improved later.
Add functions [dump_basic] and [dump_terminator] to implement
"-dcfg" compiler flag.
Keep the functions [print_basic] and [print_terminator] for printing
in dot format to a file.