Do not cache young_limit in a processor register (upstream PR 9876) (#315)
On target architectures with 32 or more registers,
a register was used to cache the value of the young_limit field
of the domain state.  This reduced the size and execution time
of the code for inlined allocations.

However, this usage is problematic with respect to polling for signals
and to inter-domain communication in Multicore OCaml, because it is
often not possible to change the value of the register when we change
young_limit.  So, the change to young_limit doesn't take effect
immediately, only when the register is reloaded from young_limit.

  - Removes the caching of young_limit in a register from the
    ARM64, PowerPC and RISC-V ports.

  - Recycle the former "young limit" register, giving one more allocatable register

    Now that we have a unused callee-save register on ARM64, PowerPC, and RISC-V,
    make it available for register allocation.

  - Assorted cleanups in runtime/*.S and in asmcomp/*/proc.ml

  - ARM64: wrong register pressure limits for Iextcall
    There are only 7 callee-save integer registers (x19 to x25), not 10.

Co-authored-by: Xavier Leroy <xavier.leroy@college-de-france.fr>