Import flexpect from old repo and add to Makefile/dune (#340)
You can now `make runtest` to run a couple of tests. One of them is an
`flexpect`-based test, with a functioning `make promote`. (The other
is the preexisting `meet_test.ml`, which simply needed to be made a
test in `dune`.)

There are also the `fldiff` and `parseflambda` utilities from the old
repo. `fldiff` might be useful for testing the `Flambda2_compare`
module, and `parseflambda` more or less runs the middle end and
nothing else.