Simplify checking whether flambda2 is enabled in Makefile
Currently we have a rather roundabout way of figuring out whether flambda2 is
enabled when we run the upstream tests. Since the Makefile is getting generated
by autoconf anyway, we can just check a Make-level variable instead.

Note that the `middle_end` variable is currently completely unused, both in
`configure.ac` and in `Makefile.in`, so changing its value here is safe.

(It would further improve things to use a Make-level conditional rather than a
shell `if` to make for better log output, but the combined foibles of Make and
shell syntax make it hard to find a not-horrendous way to do this.)