Switch the implementation of out of bounds (#586)
* Switch the implementation of out of bounds

Previously, the raising of the out of bounds exn caled
into a C function. That was mainly because at the time,
we could not lift exception (iirc it was before the
introduction of the Immutable unique thing, so exn were
declared as mutable to avoid copying).

Now that exceptions are lifted as expected, we can restore
the code path that uses apply_cont to raise the exception.

Also: the old code was mistakenly using `~alloc:false` for the
extcall, which meant that there were no backtraces for these
exceptions, so this commit also restores backtraces for
out of bounds.