Provers: split functions between prove_* and check_* versions
The prove_* versions will return a Proof when the inputs type strictly
matches the property in question, while the check_* functions will return
a Known_result when the corresponding operation has a result that can be
easily extracted.

As an example, for untagging the `prove_tagging_of_simple` version will
return Unknown on option-like variants, while the `check_tagging_of_simple`
version will assume the block part of the type is irrelevant.