Refactor Patricia_tree to reduce duplicated code
We use Patricia trees to implement both sets and maps. It's tempting then to
define `Set.t` as `unit Map.t`, but this wastes a word for every member of every
set.

This patch accomplishes much the same thing, without the storage
overhead. The idea is to write most of the operations inside a
functor that abstracts over the specific tree datatype. With a bit
of care, this can be done so that all the abstraction inlines away.

(There is some overhead when compiled with Closure due to its limited inliner,
but mostly only with sets (because `Map0.descr` is the identity but `Set0.descr`
allocates). Avoiding further allocations largely requires either (a) major
surgery on the algorithmic parts or (b) either local allocations or unboxed
types so that returning a view is cheap.)

Besides the obvious savings in repeated code, this should also make it easy to
write efficient mixed set/map operations (like restricting a map to the keys in
a set in O(n) time).

I've confirmed that the overhead is only ~2 seconds (out of ~100) when compiling
the stage1 compiler (via Closure), and negligible when compiled with
flambda2. (A few functions gain extra arguments that are always zero.)

There's a lot of changed code here, but most of the sensitive parts of
`Make_map` are unchanged but for some extra syntactic noise.