Add Minicheck library and use it to test `Patricia_tree`
Minicheck is a quick-and-dirty knockoff of Quickcheck that should be enough to
do meaningful tests of the Patricia tree implementation. The checks I've written
should be thorough, providing complete specifications for each function. (For
instance, `A` and `B` are both subsets of `union A B`, and everything in
`A union B` is in either `A` or `B`.)