Support large frames in Emitaux
Check size when frame is recorded not when it is emitted.
Move checks from emit_frame to `record_frame_descr` to give a more
meaningful error message with a function name and
record whether the frame is long.

When compiled with -allow-long-frames, emit the long frames frame
using 32 bit (whereas we use 16 bit constants for short frames),
intead of raising user error.