Improved colour API with nested colour regions
This turns all the colour functions like `Flambda_colours.error` into
functions of type `Format.formatter -> unit`, to be passed to the
`%t` format specifier in `fprintf`. This bakes in the `@<0>` adjustment
that makes sure the ANSI sequence doesn't cause excessive line wrapping
(this adjustment was missing in several places before). In addition, the
new `Flambda_colours.pop` function restores the previous colour state by
popping from a stack. Previously, any code that applied a colour had to
know what the previous colour was in order to revert; this way is easier
and potentially more modular. (Also, it should now be possible to add a
background colour to a whole region of code if we want.)

The one downside is that we need to be sure that every non-`pop` colour
directive is matched by a `pop`. This, however, is no harder than
keeping `@[` specifiers balanced.

Also, the state stack is kept in a global. It might be better to use
`Format`'s semantic tags, which allow keeping per-formatter state.