Fix soundness bug by using currying information from typing (#850)
Previously, transl_curried_function in Translcore redetected function
currying, which is difficult with locals as modes may make it invalid
to merge two lambdas into a single n-ary function.

The mode logic here was wrong, leading to a soundness bug. Rather than
fix it (which would continue the duplication of mode-checking between
typing and transl), the fix here is to add the relevant information to
Typedtree, so that Translcore follows the decisions made by typing
instead of redetecting currying on its own.