Add a flag to regions to control `Iendregion` elision
Currently, `selectgen.ml` omits the `Iendregion` instruction whenever it
compiles a `tail` expression that is not in (function-level) tail position. This
is safe under current assumptions, but a forthcoming PR will lift `region`
expressions and, in doing so, move arbitrary code into `tail` expressions. A
lifted `region` must therefore not omit the `Iendregion` when compiling a
`tail`. In return, such a region will always _have_ a `tail` expression (or
close-at-apply call), so `selectgen.ml` can compile a `region` expression
without proactively _adding_ an `Iendregion`.

This adds a `tail_policy` flag to each `region` expression to control which
behavior is required. The flag has values `May_drop_tail`, meaning there may or
may not be a `tail` and the `Iendregion` may sometimes be elided from `tail`s;
and `Must_keep_tail`, meaning there will be a `tail` and it must generate an
`Iendregion`.

Two other small updates:

* Fixed a minor bug in `inline_and_simplify.ml` (by way of `flambda.ml`) that
  sometimes causes regions to be kept in place but otherwise ignored. Currently
  this is largely benign (at worst it will cause an unused region to be kept)
  but the lifting PR will need regions to be tracked precisely.

* Undid a formatting change in hand-synched `backend` code because life is
  short and conflicts are time-consuming.