Atomic arithmetic intrinsics (#1011)
Adds builtins for the equivalent of GCC's __sync_fetch_and_add/sub and __sync_bool_compare_and_swap.

The pointer type may be native_pointer, ext_pointer, or bigstring with offset. The integer type may be untagged int, unboxed int64, unboxed int32, or unboxed nativeint.

The builtins are not translated on ARM (but it would now be easy to add ARM backend support - armv8 provides single instructions for each of these operations).

The unboxed int64 builtins are not translated on 32 bit architectures.