Add `-no-output-cmi-for-missing-mli`
This new option controls whether a `.cmi` is output in the case that we're
compiling a `.ml` with no `.mli`. Handy when compiling in both bytecode and
native so that we don't generate the same `.cmi` twice (and in the case of
packs, we're apparently generating *different* `.cmi` files, possibly only due
to differences in sharing in marshaled values).

Can't currently be combined with both `-pack` and `-bin-annot` simultaneously.
This can probably be addressed if someone needs it.