Disallow interdependencies between parameter types for now
The way they're handled now is both finicky and suspect. Given `a.mli` and
`b.mli` where `B` refers to `A`, we compile the reference to `A` as a
persistent name then substitute it away for a local name later. And the
fact that `A` is locally bound in `B` is implicit during compilation of `B`,
unlike other cases of references to parameters. Soon we hope to be able to
declare that `A` simply occurs free in `B`, which fixes both problems by making
`A` a local identifier in `B` to begin with.