Fix bug in arguments to `get_unit_info`
The `cmx_name` parameter came out of a misunderstanding of what
`get_unit_info` in `compilenv.ml` needs to do. `get_unit_info` currently takes
the compilation unit that someone intended to access, and the .cmx file that
contains that unit (possibly as part of a pack). The problem is that we expect
the .cmx's `ui_unit` to be the name _of the pack_, and similarly the import that
we record should be the import _of the pack_. So we don't actually care about
the pack member, just the .cmx. We do need the full compilation unit of the
.cmx, not just its name, so that compilation unit is now the sole argument to
`get_unit_info`.

This required changing `Compilation_unit.which_cmx_file` to return the full path
rather than just the name.