Refactoring and fixes around module lookup (#107)
* Don't invoke strengthening directly in Env

This is a refactoring commit with not functional changes. The aim is to
move all knowledge about strengthening from Env to Mtype where it arguably
belongs. The changes are:

  - `scrape_alias moves` from Env to Mtype
  - the functionality of `find_strengthened_module` also moves from Env to
    Mtype but with a slightly different interface (cf.
    `find_type_of_module`)
  - some knock-one refactoring and simplifications.

* Fix bug in includemode_errorprinter

This was using `find_modtype` on aliases which is just wrong.

* Refactor destructive substitution

This just pulls out some (very little) common code in preparation for
the more complex things we'll have to do here once we start delaying
`with` constraints.

* Refactor contains_type

This is just a small simplification.

* Simplify alias parameter in find_module*

As suggested by @lpw25, this removes the `alias` parameter from
`find_module` and doesn't expose it from `find_module_lazy` (effectively
reverting part of a previous change). For the latter, it fixes a
long-standing bug in `find_type_of_module` (previously
`find_strengthened_module`).

---------

Co-authored-by: Roman Leshchinskiy <rleshchinskiy@janestreet.com>