Support for variant kinds in .fl syntax
The syntax is meant to be reminiscent of polymorphic variants, so
`[ 0 | 1 | 0 of imm tagged | 1 of float boxed * int64 boxed ]` for a variant
with two constants and two non-constant constructors.

Also includes a new roundtrip test that we can use to make sure that the parser
and printer don't accidentally go out of sync with the language. It needs
work (something to generate dune targets, roundtripping .flt files, and a test
to generate an .flt and roundtrip _that_) but it's a start.