Add parameters to compilation units
A parameter in this case is simply a module name that is allowed to occur free.
Its type is determined by an .mli named after the parameter. This parameter must
be compiled using `-as-parameter` and cannot be implemented directly.

Currently, there is no way to use a module with a parameter. For that, we need
the `-instantiate` operation, which will take the module and substitute a real
module in place of its parameters.

Under the hood, a parameter is just a functor parameter. Using open terms rather
than functors will make things compose much more easily: rather than needing to
deal with the messy types that interdependent functors entail, the user will
simply be able to refer to one parameterised module from another.