Generated Dune rules for generating, regenerating tests (#1150)
* Automatically generate Dune files for tests in `mlexamples`

* Improve CLI for dumping .fl, .flt files

This introduces `-drawfexpr-to` and `-dfexpr-to`, which dump an `.fl` file
before or after the simplifier runs, respectively. Also `-dflexpect` is now
`-dflexpect-to` and takes the filename as an option.

The new CLI will make it very easy to regenerate flexpect tests and
roundtripping tests from Dune.

* Generated Dune rules for generating, regenerating tests

This will make adding lots of tests from `mlexamples` to CI as easy as adding a
single basename to a list in `mlexamples/dune`. An `flexpect` test is generated
along with roundtrip tests for both the pre- and post-simplifier code.