Memoise last substitution composition
This adds memoisation for the last `compose` operation to `Subst`. It
seems that we often unfold a lazy signature by one level and then
substitute into it. In such a case, we will repeatedly compose the same
two substitutions for each lazy child of the signature. By memoising
this last composition, we avoid this extra work which seems to be worth
about 1% in both time and allocations currently. This will become much
more important when we introduce lazy value_descriptions.

It doesn't seem to matter much which of the two substitutions we memoise
on. I also tried just memoising one composition globally and that seemed
to be slightly worse.