`make install` puts ocamlc at workspace root (#152)
This `ocamlc` is nicer to use with `ocamldebug`, as
the `ocamldebug` session inherits the pwd of the bytecode
file you pass as argument. In particular, breakpoints bring
you to the proper file if you use the root `ocamlc`, and
bring you to the copied file in the `_build` directory if
you use the `$USER-out/bin/` `ocamlc`.