Preserve order of declarations in various places
Having things in predictable order is important for `flexpect` to work because
the comparison engine relies on declarations appearing in the same order between
the actual and expected output. I found a few places where declarations are
produced by folding over a `Map` where we have (or can make) a list or `Lmap` to
use instead. Also, there are places where a list or `Lmap` is folded over and
rebuilt; these could probably be fixed by adding a `List.rev` but it's clearer
just to use an order-preserving operation like `map` or `fold_left_map`, so I've
done so.