Give `my_closure` accurate rec info
Since `my_closure` is always the same value as the name being defined, we give
it an alias type. Unfortunately, though they are the same runtime values, they
have different rec info: `my_closure` always has depth `my_depth`. The solution
is simply to add a coercion to the type of `my_closure`, giving us (if defininig
a function called `foo`):

    my_closure = foo ~ change_depth 0 -> my_depth

This is particularly important for lifted functions, since in that case we will
actually rewrite references to `my_closure` to use the symbol, and previously
this was causing ill-typed references:

    apply (my_closure ~ change_depth my_depth -> succ my_depth) ...

gets rewritten to

    apply ($Foo ~ change_depth my_depth -> succ my_depth) ...

This coercion application is ill-typed since the symbol `$Foo` always has depth
zero. If `my_closure` carries the proper coercion, however, we get instead

    apply (($Foo ~ change_depth 0 -> my_depth)
           ~ change_depth my_depth -> succ my_depth)

which simplifies to

    apply ($Foo ~ change_depth 0 -> succ my_depth)

which is correct.

I don't know of any user-visible misbehavior that would be caused by this
currently, but it can't be good to have `my_closure` carry the wrong rec info.