flambda-backend: Merge Types.* and Subst.Lazy.* types (#1312)
* Unify Types.* and Subst.Lazy.* types

We introduce a functor which generates the types instead of duplicating
them. The main benefit (in addition to removing code duplication) is
that we abstract over things which can be lazy via the Pod mechanism.
Going forward, this will allow us to make more thing lazy (esp. the
`type_expr` in `value_description` in the next set of changes) but also
to reuse the same mechanism to deserialise cmi files lazily.

---------

Co-authored-by: Roman Leshchinskiy <rleshchinskiy@janestreet.com>