Provide an AST-like view of modular extension extension node names
That is: if you have `[%extension.comprehensions.for.in]`, this will be
represented in OCaml as
`Extensions_parsing.Extension_node_name.("comprehensions" :: ["for"; "in"])`;
we're guaranteed to have it be nonempty, we can pass it around as a unit, and
most importantly we can change the representation in a single place (e.g., to
`[%jst.comprehensions.for.in]`, or to `[%extension'comprehensions'for'in]`,
etc.).

Some minor error changes come along as well, either around spacing (due to
printing changes) or the locations in errors that the user will never see.