Don't substitute into an exclave in `un_anf.ml`
In cases like

    let local_ n : nativeint = ... in
    let i = Nativeint.to_int n in
    exclave (
      let local_ a = ... in
      ... i ...
    )

we don't want to substitute `i`, since that would read from `n` outside of
its scope. In particular, `a` here will be allocated over `n`, so
`Nativeint.to_int` will read garbage.

Simple solution is to treat an `exclave` like a loop during the analysis that
decides whether a variable is used linearly.