Fix max indent
Currently the output of `-dlambda` and such can easily become unreadable at
large nesting depths, even if you set `COLUMNS=200`, because we set the
pretty-printing /width/ equal to `COLUMNS` (minus a bit of slop) but the default
/maximal indentation level/ doesn't scale with the width, so what should be
nested elements end up left-aligned with each other, making the nesting
structure impossible to see.

This sets the indentation level to be at least 75% of the max width, ensuring
that we keep indenting at least as long as the remaining line is 25% of the
screen.

I picked 75% somewhat arbitrarily. Notably, the default ratio is higher:

    $ ocaml
    OCaml version 4.14.0
    Enter #help;; for help.

    # Format.pp_get_margin Format.std_formatter ();;
    - : int = 78
    # Format.pp_get_max_indent Format.std_formatter ();;
    - : int = 68