Flambda1: Tolerate removing all closures from a set
In rare circumstances, we lift a set of closures into a symbol and later find
that none of its closures is actually used. Since we don't remove dead
`Initialize_symbol`s, the symbol will remain, but since all the closure ids are
dead, we'll try to remove _all_ of the closures from the set. This crashes,
since we don't actually allow an empty set of closures. However, we know that
all the closure ids are dead, which means that no one is actually observing that
the symbol is bound to a set of closures. Therefore we can simply bind the
symbol to a constant zero instead.