Delete sort_let_bound
This deletes a couple safety checks in transl_core.  This is fine, because these
checks were there to guard against calling [transl_exp] with an expression that
has a non-value type (and [transl_exp] now supports that).  We still hit an
safety check error in this case, it just comes later - these are the changes
where errors that didn't used to mention [value_kind] now do.

In subsequent commits these may change again to mention [layout], because we'll
be providing accurate sorts to [layout] rather than the pre-defined ones in
various places.