Don't substitute into exclaves in `simplif.ml`
The current behavior in the lambda-to-lambda optimizer substitutes away certain
`let` bindings that are used only once and not under a lambda or inside a loop.
We need to add `exclave` to this list of provisos, to avoid cases like:

```
type 'a glob = Glob of ('a[@global])

let[@inline never] return_local a = [%local] (Glob a)

let f () =
  let (Glob x) = return_local 1 in
  [%exclave]
    (let (_ : _) = return_local 99 in
     assert (x <> 99))
;;
```

Here the lambda code generated is something like

```
let *match* = apply return_local 1 in
let x = field 0 *match* in
exclave (
  let _ = apply return_local 99 in
  assert (x <> 99)
)
```

If we read `field 0 *match*` inside the exclave, we'll get 99 rather than 1,
since the exclave decrements the local stack pointer, meaning the `Glob 99`
block gets allocated on top of the `Glob 1`, clobbering the 1 with a 99.
Therefore we have to keep the binding of `x` where it is. Conservatively, we
assume that _any_ `let` binding could be unsafe to move into the `exclave`, so
we treat `exclave` like a loop body and never substitute into it.