Consult CRC table directly in `crc_of_unit`
This appears to be the other place we need the code that I want to delete. But
it doesn't _actually_ need the named unit to be in the table of persistent
structures - if it finds the persistent structure, it only uses it to find the
CRC, which we already know in the `penv` itself.