Substitute dependencies as well as arguments
Just as dependencies become runtime arguments, when importing a module we need
to substitute dependencies along with arguments. This requires exporting in the
.cmi the list of globals that the module's type depends on (which is then also
a good place to cache the global for each name).

Also added a test (though "basic" is perhaps a misnomer) that covers forwarding
of arguments by subset rule as well as the use of instance identifiers. Not
covered (because it's not implemented in this branch) is instantiation or
anything depending on an complete application (i.e., a concrete instance).