Better region tracking in flambda1 (#1497)
Regions are currently not tracked very precisely in flambda1. In particular, it
is assumed that nothing in a region block can affect the parent region, which is
not true when there are explicit exclaves. Also, some old traversal code tends
to ignore region borders entirely, which is almost certainly causing subtle bugs
somewhere (and adding region lifting makes it cause unsubtle bugs everywhere).

Adding an actual region stack makes it easy to be precise, and will also help
remove regions that have enclaves (see upcoming PR).