Define `Signature_with_global_bindings`
This is just a pair with a module's signature and its bound global names. This
is useful because it's a complete description of the module type, and in
particular, when we need to substitute arguments into that module type, that
substitution is well-described as an operation on
`Signature_with_global_bindings.t`.