Stratify `pers_struct` and avoid binding too many names
This is partly just PR #1674, but we're going further: Not only do we avoid
adding bindings when reading .cmi files, we do so when resolving names as usual
as well when those names are parts of instance names - adding an argument value
to the environment is unnecessary, and adding an argument name to the
environment is actively bad since it can't necessarily be referred to in source
by itself.

In trunk, there is just `persistent_structures`, a hash table with an entry for
each loaded global/module/.cmi file. As of this patch, we have `imports`,
`persistent_names`, and `persistent_structures`. For any .cmi we have an entry
in `imports` that covers just the .cmi-level data: in particular, things that do
not change no matter what arguments are given. Then, for each `Global.Name.t`
(which is to say, each pair of .cmi name and list of arguments), there is an
entry in `persistent_names` that has all the type information associated with
that global name, whether or not it corresponds to anything actually bound in
the environment. Finally, each global name that _is_ bound gets an entry in
`persistent_structures`. These three caches form a sequence: each `pers_name` is
built from (and includes by reference) a cached `import`, and each `pers_struct`
is built from (and includes) a cached `pers_name`.

Other changes:

* The table of global name bindings in a .cmi is now just a `Global.t array`
  rather than a `(Global.Name.t * Global.t) array`, since the global name is
  always just the erasure of the global
* Building on PR #1670, the freshening substitution is now applied outside of
  `Env` and will now be applied only once per loaded .cmi (as the substituted
  signature will be put in the `import`)