Combine several test actions
It's a bit of a hack, but it turns out if the `module` field of an ocamltest
compilation action has multiple words, it just happily passes multiple filenames
to the command line, which just works. This cuts down on noise considerably
without losing clarity for things like compiling the .ml and .mli for the same
module with nearly identical command lines.