flambda-backend: Add code path to read .cmi without adding to environment  (#1674)
Currently, every time a .cmi is read for any reason, we bind the name of the
module in the environment. In nearly every case, this is undesirable: for
instance, if we're reading the .cmi for the current module, we do _not_ want
to add the current module to its own environment. This behavior is largely
benign at the moment since we read .cmi files after typechecking, but
parameterised libraries will complicate this picture.